Resolved warnings
