+++ src/games/strategy/engine/framework/map/download/DownloadFile.java
@@ -10,7 +10,6 @@ import com.google.common.io.Files;
 
 import games.strategy.debug.ClientLogger;
 import games.strategy.engine.ClientFileSystemHelper;
-import games.strategy.performance.Perf;
 import games.strategy.performance.PerfTimer;
 
 /**
@@ -68,14 +67,14 @@ public class DownloadFile {
   private Thread createDownloadThread(final File fileToDownloadTo) {
     return new Thread(() -> {
       if (state != DownloadState.CANCELLED) {
-        try (PerfTimer timer = Perf.startTimer("Download map: " + downloadDescription.getUrl())) {
-          URL url = downloadDescription.newURL();
-          try {
-            DownloadUtils.downloadFile(url, fileToDownloadTo);
-          } catch (Exception e) {
-            ClientLogger.logError("Failed to download: " + url, e);
-          }
+        PerfTimer timer = PerfTimer.startTimer("Download map: " + downloadDescription.getUrl());
+        URL url = downloadDescription.newURL();
+        try {
+          DownloadUtils.downloadFile(url, fileToDownloadTo);
+        } catch (Exception e) {
+          ClientLogger.logError("Failed to download: " + url, e);
         }
+        timer.stop();
         if (state == DownloadState.CANCELLED) {
           return;
         }
