+++ src/games/strategy/engine/history/SerializationWriter.java
@@ -1,0 +1,7 @@
+package games.strategy.engine.history;
+
+import java.io.Serializable;
+
+public interface SerializationWriter extends Serializable {
+  void write(HistoryWriter writer);
+}
