+++ src/games/strategy/engine/message/UnifiedMessenger.java
@@ -1,9 +1,5 @@
 package games.strategy.engine.message;
 
-import java.io.Externalizable;
-import java.io.IOException;
-import java.io.ObjectInput;
-import java.io.ObjectOutput;
 import java.io.PrintStream;
 import java.io.Serializable;
 import java.lang.reflect.InvocationTargetException;
@@ -575,120 +571,5 @@ class EndPointDestroyed implements Serializable {
   @Override
   public String toString() {
     return "EndPointDestroyed:" + name;
-  }
-}
-
-
-// someone now has an implementor for an endpoint
-class HasEndPointImplementor implements Serializable {
-  private static final long serialVersionUID = 7607319129099694815L;
-  public final String endPointName;
-
-  public HasEndPointImplementor(final String endPointName) {
-    this.endPointName = endPointName;
-  }
-
-  @Override
-  public String toString() {
-    return this.getClass().getName() + ":" + endPointName;
-  }
-}
-
-
-// someone no longer has implementors for an endpoint
-class NoLongerHasEndPointImplementor implements Serializable {
-  private static final long serialVersionUID = -4855990132007435355L;
-  public final String endPointName;
-
-  public NoLongerHasEndPointImplementor(final String endPointName) {
-    this.endPointName = endPointName;
-  }
-}
-
-
-// someone wants us to invoke something locally
-abstract class Invoke implements Externalizable {
-  private static final long serialVersionUID = -5453883962199970896L;
-  public GUID methodCallID;
-  public boolean needReturnValues;
-  public RemoteMethodCall call;
-
-  public Invoke() {}
-
-  @Override
-  public String toString() {
-    return "invoke on:" + call.getRemoteName() + " method name:" + call.getMethodName() + " method call id:"
-        + methodCallID;
-  }
-
-  public Invoke(final GUID methodCallID, final boolean needReturnValues, final RemoteMethodCall call) {
-    if (needReturnValues && methodCallID == null) {
-      throw new IllegalArgumentException("Cant have no id and need return values");
-    }
-    if (!needReturnValues && methodCallID != null) {
-      throw new IllegalArgumentException("Cant have id and not need return values");
-    }
-    this.methodCallID = methodCallID;
-    this.needReturnValues = needReturnValues;
-    this.call = call;
-  }
-
-  @Override
-  public void readExternal(final ObjectInput in) throws IOException, ClassNotFoundException {
-    needReturnValues = in.read() == 1;
-    if (needReturnValues) {
-      methodCallID = (GUID) in.readObject();
-    }
-    call = new RemoteMethodCall();
-    call.readExternal(in);
-  }
-
-  @Override
-  public void writeExternal(final ObjectOutput out) throws IOException {
-    out.write(needReturnValues ? 1 : 0);
-    if (needReturnValues) {
-      out.writeObject(methodCallID);
-    }
-    call.writeExternal(out);
-  }
-}
-
-
-// the results of a remote invocation
-abstract class InvocationResults implements Externalizable {
-  private static final long serialVersionUID = -382704036681832123L;
-  public RemoteMethodCallResults results;
-  public GUID methodCallID;
-
-  public InvocationResults() {}
-
-  public InvocationResults(final RemoteMethodCallResults results, final GUID methodCallID) {
-    if (results == null) {
-      throw new IllegalArgumentException("Null results");
-    }
-    if (methodCallID == null) {
-      throw new IllegalArgumentException("Null id");
-    }
-    this.results = results;
-    this.methodCallID = methodCallID;
-  }
-
-  @Override
-  public String toString() {
-    return "Invocation results for method id:" + methodCallID + " results:" + results;
-  }
-
-  @Override
-  public void writeExternal(final ObjectOutput out) throws IOException {
-    results.writeExternal(out);
-    methodCallID.writeExternal(out);
-  }
-
-  @Override
-  public void readExternal(final ObjectInput in) throws IOException, ClassNotFoundException {
-    results = new RemoteMethodCallResults();
-    results.readExternal(in);
-    methodCallID = new GUID();
-    methodCallID.readExternal(in);
   }
 }
